Android GUI: Split patcher options loading to LiveData subclass
Also fixes an issue where loadData wouldn't be called after the Fragment
is recreated following a out-of-memory kill.
